consolidate all StockTransaction signal receivers into one file
