+++ corehq/apps/commtrack/models.py
@@ -8,16 +8,15 @@ from couchdbkit.exceptions import ResourceNotFound
 from django.utils.translation import ugettext as _
 from django.dispatch import receiver
 from django.db import models
-from django.db.models.signals import post_save, post_delete
+from django.db.models.signals import post_save
 
 from corehq.apps.commtrack.dbaccessors import get_supply_point_case_by_location
 
 from dimagi.ext.couchdbkit import *
 from casexml.apps.case.mock import CaseBlock
 from casexml.apps.case.models import CommCareCase
-from casexml.apps.stock.consumption import (ConsumptionConfiguration, compute_default_monthly_consumption,
-    compute_daily_consumption)
-from casexml.apps.stock.models import StockReport, StockTransaction, DocDomainMapping
+from casexml.apps.stock.consumption import (ConsumptionConfiguration, compute_default_monthly_consumption)
+from casexml.apps.stock.models import StockReport, DocDomainMapping
 from casexml.apps.case.xml import V2
 from corehq.apps.cachehq.mixins import CachedCouchDocumentMixin
 from corehq.apps.commtrack import const
@@ -731,99 +730,6 @@ def sync_location_supply_point(loc):
             sql_loc.save()
         except SQLLocation.DoesNotExist:
             pass
-
-
-@receiver(post_save, sender=StockTransaction)
-def update_stock_state_signal_catcher(sender, instance, *args, **kwargs):
-    update_stock_state_for_transaction(instance)
-
-
-def update_stock_state_for_transaction(instance):
-    # todo: in the worst case, this function makes
-    # - three calls to couch (for the case, domain, and commtrack config)
-    # - four postgres queries (transacitons, product, location, and state)
-    # - one postgres write (to save the state)
-    # and that doesn't even include the consumption calc, which can do a whole
-    # bunch more work and hit the database.
-    try:
-        domain_name = instance.domain
-    except AttributeError:
-        domain_name = CommCareCase.get(instance.case_id).domain
-
-    domain = Domain.get_by_name(domain_name)
-
-    sql_product = SQLProduct.objects.get(product_id=instance.product_id)
-
-    try:
-        sql_location = SQLLocation.objects.get(supply_point_id=instance.case_id)
-    except SQLLocation.DoesNotExist:
-        sql_location = None
-
-    try:
-        state = StockState.include_archived.get(
-            section_id=instance.section_id,
-            case_id=instance.case_id,
-            product_id=instance.product_id,
-        )
-    except StockState.DoesNotExist:
-        state = StockState(
-            section_id=instance.section_id,
-            case_id=instance.case_id,
-            product_id=instance.product_id,
-            sql_product=sql_product,
-            sql_location=sql_location,
-        )
-
-    # we may not be saving the latest transaction so make sure we use that
-    # todo: this should change to server date
-    latest_transaction = StockTransaction.latest(
-        case_id=instance.case_id,
-        section_id=instance.section_id,
-        product_id=instance.product_id
-    )
-    if latest_transaction != instance:
-        logging.warning(
-            'Just fired signal for a stale stock transaction. Domain: {}, instance: {},latest was {}'.format(
-                domain_name, instance, latest_transaction
-            )
-        )
-        instance = latest_transaction
-    state.last_modified_date = instance.report.date
-    state.stock_on_hand = instance.stock_on_hand
-
-    if domain and domain.commtrack_settings:
-        consumption_calc = domain.commtrack_settings.get_consumption_config()
-    else:
-        consumption_calc = None
-
-    state.daily_consumption = compute_daily_consumption(
-        instance.case_id,
-        instance.product_id,
-        instance.report.date,
-        'stock',
-        consumption_calc
-    )
-    # so you don't have to look it up again in the signal receivers
-    if domain:
-        state.domain = domain.name
-    state.save()
-
-
-@receiver(post_delete, sender=StockTransaction)
-def stock_state_deleted(sender, instance, *args, **kwargs):
-    qs = StockTransaction.objects.filter(
-        section_id=instance.section_id,
-        case_id=instance.case_id,
-        product_id=instance.product_id,
-    ).order_by('-report__date')
-    if qs:
-        update_stock_state_for_transaction(qs[0])
-    else:
-        StockState.objects.filter(
-            section_id=instance.section_id,
-            case_id=instance.case_id,
-            product_id=instance.product_id,
-        ).delete()
 
 
 @receiver(post_save, sender=StockState)
