--- corehq/ex-submodules/casexml/apps/stock/__init__.py
@@ -1,3 +1,0 @@
-
-
-
